+++ lib/ash_authentication/token_resource.ex
@@ -57,6 +57,13 @@ defmodule AshAuthentication.TokenResource do
               """,
               default: :revoke_token
             ],
+            revoke_all_stored_for_subject_action_name: [
+              type: :atom,
+              doc: """
+              The name of the action used to revoke all stored tokens for a given subject.
+              """,
+              default: :revoke_all_stored_for_subject
+            ],
             is_revoked_action_name: [
               type: :atom,
               doc: """
@@ -145,7 +152,8 @@ defmodule AshAuthentication.TokenResource do
 
   use Spark.Dsl.Extension,
     sections: @dsl,
-    transformers: [TokenResource.Transformer, TokenResource.Verifier]
+    transformers: [TokenResource.Transformer],
+    verifiers: [TokenResource.Verifier]
 
   @doc """
   Has the token been revoked?